+++ client/src/lib/externalmodules.ts
@@ -1,6 +1,6 @@
 // keytar depends on a native module shipped in vscode
 // this loads only the type definitions
-import * as keytarType from "keytar"
+import * as keytarType from "./keytar"
 
 // get the module from vscode. This is not an official API, might break at some point
 // this is required because keytar includes a binary we can't include
@@ -25,16 +25,24 @@ export class PasswordVault {
   constructor() {
     if (!keytar) keytar = getCodeModule<typeof keytarType>("keytar")
   }
+
   getPassword(service: string, account: string) {
     if (!keytar) throw keytarErr()
     return keytar.getPassword(service, account)
   }
+
   setPassword(service: string, account: string, password: string) {
     if (!keytar) throw keytarErr()
     return keytar.setPassword(service, account, password)
   }
+
   deletePassword(service: string, account: string) {
     if (!keytar) throw keytarErr()
     return keytar.deletePassword(service, account)
+  }
+
+  acounts(service: string) {
+    if (!keytar) throw keytarErr()
+    return keytar.findCredentials(service)
   }
 }